+++ quickmlops/quickmlops/templates/scikit_learn/train.py
